pyproject.toml
[project]
name = "bloqade-circuit"
version = "0.8.0-DEV"
description = "The software development toolkit for neutral atom arrays."
readme = "README.md"
authors = [
    { name = "Roger-luo", email = "rluo@quera.com" },
    { name = "kaihsin", email="khwu@quera.com" },
    { name = "weinbe58", email="pweinberg@quera.com"},
    { name = "johnzl-777", email="jlong@quera.com"},
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.22.0",
    "scipy>=1.13.1",
    "kirin-toolchain~=0.17.30",
    "rich>=13.9.4",
    "pydantic>=1.3.0,<2.11.0",
    "pandas>=2.2.3",
    "pyqrack-cpu>=1.69.1",
]

[project.optional-dependencies]
qasm2 = [
    "lark>=1.2.2",
]
vis = [
    "tqdm>=4.66.5",
    "matplotlib>=3.9.2",
    "pyqt5>=5.15.11;  sys_platform == 'darwin'",
    "pyqt5>=5.15.11;  sys_platform == 'linux'",
    "ffmpeg>=1.4",
]
qbraid = [
    "qbraid>=0.9.3",
]
cirq = [
    "cirq-core>=1.4.1",
    "cirq-core[contrib]>=1.4.1",
    "qpsolvers[clarabel]>=4.7.0",
]
pyqrack-opencl = [
    "pyqrack>=1.69.1",
]
pyqrack-cuda = [
    "pyqrack-cuda>=1.69.1",
]
stim = [
    "stim>=1.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bloqade"]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "coverage>=7.6.4",
    "ipython>=8.29.0",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "numpy>=1.26.4",
    "pre-commit>=4.0.1",
    "pyright>=1.1.388",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
    "rust-just>=1.36.0",
    "tomlkit>=0.13.2",
]

[tool.isort]
profile = "black"
combine_as_imports = true
multi_line_output = 3
length_sort = true
src_paths = ["src/bloqade"]

[tool.black]
line-length = 88

[tool.ruff]
target-version = "py310"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.coverage.run]
include = ["src/bloqade/*"]

[tool.pytest.ini_options]
testpaths = "test/"
filterwarnings = [
    "ignore:In cirq 1.6 the default value of `use_repetition_ids`:FutureWarning:cirq.circuits.circuit_operation",
]
